Fixed: Incorrect usage of ZStandard API
Changes the code to correct the error handling for: ZSTD_decompressStream

While the original code is correct, for the cases of decompressing entire blocks; it's not valid for the case of partial decompression; because the result of ZSTD_decompressStream can return a non-error positive value that's used as a 'hint' for the next call to `ZSTD_decompressStream`. 

Previously we checked the result against 0, which is only true if the entire input has been decompressed; but in the case of partial decompression, this cannot be true.